application/Grounded-sam/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.0.3
Pillow==10.4.0
pycocotools==2.0.7
Requests==2.32.3
sophon==3.8.0
streamlit==1.36.0
torch==2.3.1
torchvision==0.18.1
transformers==4.42.3
